backend/requirements.txt
# DRIFT backend — merged deps from Phase 1/2/3 + team's main
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.6.4
pydantic-settings==2.2.1
PyYAML>=6.0
shapely==2.0.3
geopandas==0.14.3
fiona<1.10
xarray>=2025.0
netcdf4==1.6.5
numpy==1.26.4
pandas>=2.2,<3.0
pyproj==3.6.1
scikit-learn==1.4.1.post1
scikit-image==0.22.0
utm==0.7.0
copernicusmarine>=2.3,<3.0
cdsapi>=0.7
h5py>=3.10
zarr<3
pytest==8.1.1
reportlab>=4.2,<5.0
httpx>=0.26,<0.27
matplotlib>=3.9,<4.0

# ML runtime (required by backend/ml/*)
torch==2.7.0
torchvision==0.22.0
segmentation-models-pytorch>=0.5.0,<0.6.0

# --- from team's main branch (additions not in rescue) ---
python-dotenv==1.0.1
pystac-client>=0.7.5
pystac>=1.9.0
rasterio>=1.3.10
rioxarray>=0.15.1
global_land_mask>=1.0.0
geojson-pydantic>=1.0.0
pyogrio>=0.7
requests>=2.31
